Main Application/requirements.txt
grpcio-tools >= 1.30
psycopg2-binary>=2.9
SQLAlchemy>=1.4.23
alembic>=1.6.5
ConfigParser>=5.2.0
sqlalchemy-filters>=0.12.0